requirements.txt
awsebcli==3.12.3
botocore==1.9.2
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
Django==2.0.2
django-filter==1.1.0
django-leaflet==0.23.0
djangorestframework==3.7.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
httpie==0.9.9
idna==2.6
jmespath==0.9.3
Markdown==2.6.11
pathspec==0.5.5
psycopg2==2.7.4
psycopg2-binary==2.7.4
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
websocket-client==0.47.0
